Split tests for poly_fmult* into several files
